backend/requirements.txt
# Backend dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.0

# Background jobs
arq>=0.25.0
redis>=5.0.0

# Chess
python-chess>=1.10.0

# Auth (JWT verification for NextAuth.js tokens)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Payments
paddle-python-sdk>=1.0.0

# HTTP
httpx>=0.26.0
requests>=2.31.0

# AI Coach
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
pandas>=2.1.0
